requirements.txt
transformers==4.30.2
torch>=2.0.0
chromadb==0.4.14
numpy>=1.24.0
tqdm>=4.65.0
Pillow>=9.5.0
flask==2.3.3
flask-cors==4.0.0
streamlit==1.24.1
requests>=2.31.0